requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
python-binance
binance-connector